+++ lib/crossbeams/layout/renderer/base.rb
@@ -4,7 +4,7 @@ module Crossbeams
   module Layout
     module Renderer
       # Base class for all field renderers.
-      class Base
+      class Base # rubocop:disable Metrics/ClassLength
         # Create reasonable label text from a field name.
         def present_field_as_label(field)
           field.to_s.sub(/_id$/, '').split('_').map(&:capitalize).join(' ')