+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 ### Changed
+- Styling of fold_ups improved.
+- Expand/Collapse can be called within a Column.
 ### Fixed
 
 ## [0.3.4] - 2019-09-13